#112 [refactor] companion object 중복사용 방지하기 위해 한 곳에 적용
